Introduce variations in analysis results
Today's results include one best move, one score and that's it, regardless of
what MultiPv is configured. This makes it less than ideal for external software
to build on more complex features.

This changes that while keeping backwards compatibility. Now, besides best move
and score, there's a new `variations` key in the analysis results containing an
array of the multiple principal variations (MultiPv), each including a variation
with N depth moves, as well as score, time for processing and score type.

The challenge behind this change was that each line in the UCI multiline output
is processed separatelly, and no state is kept in between them. In order to make
it possible to return the correct MultiPv, the one with the greatest depth, I
introduced an analysis cache in the `Handler` object, that is inject into the
new output analysis function, adding state in between output lines processing.
When analysis is complete, `handler.reset()` takes care of cleaning that cache
for subsequent calls to take place.

Additionally, this cache is duplicated to avoid leaking state in between tests and
instances. More specifically, `parsePrincipalVariations` runs
`candidates = Object.assign({}, candidates)` to ensure this method doesn't
modify the reference being passed in. We use `Object.assign` here because we
don't care about deep cloning for now. Shallow cloning should work.